SwordServer/requirements.txt
redis==3.5.3
grpcio==1.31.0
grpcio-tools==1.31.0
protobuf==3.11.3